Feat: "메인 페이지 기능"
메인 페이지에 RoomHistory 리스트 전달

Task: S10P12D203-111

+++ Backend/tranvel/src/main/java/com/ssafy/tranvel/repository/AnnouncementRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface AnnouncementRepository extends JpaRepository<Announcement, Integer> {
+public interface AnnouncementRepository extends JpaRepository<Announcement, Long> {
 
 }
